perf(AddTorrentDialog): Deduplicate AddTorrent form and AddTorrentParamsForm
